requirements.txt
torch
torchvision
flask
pillow
numpy
tqdm
scikit-learn
